Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
psycopg2-binary
pandas
